MyTools.Ini

Types

MyTools.Ini.IniFile — Type

IniFile(fname::AbstractString)

IniFile struct, with several sections, including default section. Construct it with a file name.

MyTools.Ini.IniSection — Type

A ini file section.

MyTools.Ini.IniItem — Type

A key = value item.

Functions

MyTools.Ini.isgood — Function

isgood(ini::IniFile)

Determine if the ini file is opened successful.

MyTools.Ini.error_msg — Function

error_msg(ini::IniFile)

Get error message of IniFile.

MyTools.Ini.getValue — Function

getValue(ini::IniFile, section::AbstractString, key::AbstractString)
getValue(ini::IniFile, key::AbstractString)

Get a value with a key in one section. The version without section name searches the default section.

MyTools.Ini.getString — Function

getString(ini::IniFile, section::AbstractString, key::AbstractString)
getString(ini::IniFile, key::AbstractString)

Get a string value with a key in one section. The version without section name searches the default section.

MyTools.Ini.getBool — Function

getBool(ini::IniFile, section::AbstractString, key::AbstractString)
getBool(ini::IniFile, key::AbstractString)

Get a boolean value with a key in one section. The version without section name searches the default section.

MyTools.Ini.getInt — Function

getInt(ini::IniFile, section::AbstractString, key::AbstractString)
getInt(ini::IniFile, key::AbstractString)

Get a integer value with a key in one section. The version without section name searches the default section.

MyTools.Ini.getFloat — Function

getFloat(ini::IniFile, section::AbstractString, key::AbstractString)
getFloat(ini::IniFile, key::AbstractString)

Get a floating point value with a key in one section. The version without section name searches the default section.

Base.show — Method

Base.show(io::IO, ini::IniFile)

Import Base.show to print a IniFile.

MyTools.Ini.saveAs — Function

saveAs(ini::IniFile, fname::AbstractString)

Save the ini file to file.